[svn-r29990] Misc warning cleanup in the tests.
Tested on: 64-bit Ubuntu 16.04 LTS w/ gcc 5.3.1
           Autotools serial

+++ test/swmr_common.c
@@ -98,7 +98,7 @@ choose_dataset(void)
     level = symbol_mapping[HDrandom() % NMAPPING];
 
     /* Determine the offset of the level */
-    offset = HDrandom() % symbol_count[level];
+    offset = (unsigned)(HDrandom() % (int)symbol_count[level]);
 
     return &symbol_info[level][offset];
 } /* end choose_dataset() */
